backend/requirements.txt
Django
djangorestframework
pillow
django-cors-headers